Move tests from test_utils.py to test_parameters.py
